move commands into own namespace
