+++ src/Blockfrost.ConsoleTool/Blockfrost.ConsoleTool.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFrameworks>net5.0</TargetFrameworks>
 
     <PackAsTool>true</PackAsTool>
     <ToolCommandName>blockfrost</ToolCommandName>
@@ -23,7 +23,7 @@
     <NrtRemoveTagV>true</NrtRemoveTagV>
     <NrtRequiredVcs>git</NrtRequiredVcs>
     <NrtShowRevision>true</NrtShowRevision>
-    <AssemblyName>Blockfrost.ConsoleTool</AssemblyName>
+    <AssemblyName>Blockfrost.Cli</AssemblyName>
     <RootNamespace>Blockfrost.ConsoleTool</RootNamespace>
     <UserSecretsId>a7eb06d1-f643-4add-bc36-37eab11e9c8b</UserSecretsId>
   </PropertyGroup>
