+++ src/Blockfrost.ConsoleTool/Commands/AddressCommand.cs
@@ -1,10 +1,9 @@
 ﻿using System;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Blockfrost.Api;
 
-namespace Blockfrost.ConsoleTool
+namespace Blockfrost.ConsoleTool.Commands
 {
     public class AddressCommand : ICommand
     {
