+++ src/Blockfrost.ConsoleTool/Commands/HealthCommand.cs
@@ -4,7 +4,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Blockfrost.Api;
 
-namespace Blockfrost.ConsoleTool
+namespace Blockfrost.ConsoleTool.Commands
 {
     public class HealthCommand : BlockfrostCommandBase
     {
